+++ versions/1.15.2/src/main/java/com/ayakacraft/carpetayakaaddition/commands/address/Address.java
@@ -23,8 +23,11 @@ package com.ayakacraft.carpetayakaaddition.commands.address;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.dimension.DimensionType;
+import org.jetbrains.annotations.NotNull;
 
-public class Address {
+import java.util.Objects;
+
+public class Address implements Comparable<Address> {
 
     public static final String DESC_PLACEHOLDER = "#none";
 
@@ -36,21 +39,24 @@ public class Address {
 
     private String desc;
 
-    public Address(String id, String dim, Vec3d pos, String desc) {
+    private long weight = Long.MIN_VALUE;
+
+    public Address(String id, String dim, Vec3d pos, String desc, long weight) {
         this.id = id;
         this.dim = dim;
         this.x = pos.x;
         this.y = pos.y;
         this.z = pos.z;
         setDesc(desc);
+        this.weight = weight;
     }
 
-    public Address(String id, DimensionType dim, Vec3d pos, String desc) {
-        this(id, String.valueOf(DimensionType.getId(dim)), pos, desc);
+    public Address(String id, DimensionType dim, Vec3d pos, String desc, long weight) {
+        this(id, String.valueOf(DimensionType.getId(dim)), pos, desc, weight);
     }
 
     public Address(AddressOld old) {
-        this(old.getId(), old.getDim(), old.getPos(), old.getDesc());
+        this(old.getId(), old.getDim(), old.getPos(), old.getDesc(), Long.MIN_VALUE);
     }
 
     public String getDim() {
@@ -87,6 +93,42 @@ public class Address {
 
     public void setDesc(String desc) {
         this.desc = (desc == null || desc.isEmpty()) ? DESC_PLACEHOLDER : desc;
+    }
+
+    public long getWeight() {
+        return weight;
+    }
+
+    public void onDetailDisplayed() {
+        weight += 1;
+    }
+
+    public void onTeleportedTo() {
+        weight += 3;
+    }
+
+    @Override
+    public int compareTo(@NotNull Address o) {
+        if (this.equals(o)) {
+            return 0;
+        }
+        if (this.weight == o.weight) {
+            return this.id.compareTo(o.id);
+        }
+        return Long.compare(this.weight, o.weight);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (!(o instanceof Address)) {
+            return false;
+        }
+        Address other = (Address) o;
+        return Objects.equals(other.id, this.id)
+                && Objects.equals(other.dim, this.dim)
+                && other.getPos().equals(this.getPos())
+                && Objects.equals(other.desc, this.desc)
+                && other.weight == this.weight;
     }
 
     @Override