+++ changelog.md
@@ -1,4 +1,3 @@
 ## General
 
-- Added support for Minecraft 1.21.6
-- `forceTickPlantsReintroduce` now affects chorus flowers as well
+- Addresses now listed in natual order