+++ changelog.md
@@ -1,3 +1,4 @@
 ## General
 
-- Addresses now listed in natual order
+- Addresses now listed in the order of their weight which increases when used
+- Support for Minecraft 1.21.7